src/requirements.txt
numpy==1.23.3
pandas==1.5
keras
tensorflow==2.10.0
pillow==9.2.0
matplotlib.pyplot==3.6.1